Change data types to more compact types such as int8

+++ scripts/helper/representations.py
@@ -65,7 +65,7 @@ def raw_bagofwords(resnames):
     represents the number of occurrences of that amino acid in the string."""
     onehot = raw_onehot(resnames)
 
-    return onehot.sum(axis=0)
+    return onehot.sum(axis=0).astype('int8')
 
 
 def raw_crossed_bagofwords(cdr_resnames, target_resnames):
@@ -76,7 +76,7 @@ def raw_crossed_bagofwords(cdr_resnames, target_resnames):
     cdr_bagofwords = raw_bagofwords(cdr_resnames)
     target_bagofwords = raw_bagofwords(target_resnames)
 
-    return np.outer(cdr_bagofwords, target_bagofwords).flatten()
+    return np.outer(cdr_bagofwords, target_bagofwords).flatten().astype('int8')
 
 
 def raw_meiler(resnames):
@@ -111,7 +111,7 @@ def raw_padded_onehot_meiler(resnames, max_length):
     num_features = len(residues_order) + 7  # one-hot + extra features
 
     cdr_mat = raw_onehot_meiler(resnames)
-    cdr_mat_pad = np.zeros((max_length, num_features))
+    cdr_mat_pad = np.zeros((max_length, num_features), dtype=np.float32)
     cdr_mat_pad[:cdr_mat.shape[0], :] = cdr_mat
 
     cdr_mask = np.zeros((max_length, 1), dtype=int)
